scope to the selected company

+++ app/Http/Controllers/Api/ClientController.php
@@ -13,10 +13,18 @@ class ClientController extends Controller
     /**
      * Display a listing of the resource.
      */
+    private function company(): \App\Models\Company
+    {
+        return app('company');
+    }
+
     public function index()
     {
         return ClientResource::collection(
-            Client::query()->orderBy('id', 'desc')->paginate(10)
+            Client::query()
+                ->where('company_id', $this->company()->id)
+                ->orderBy('id', 'desc')
+                ->paginate(10)
         );
     }
 
@@ -30,6 +38,7 @@ class ClientController extends Controller
             'company_address' => 'required',
         ]);
 
+        $data['company_id'] = $this->company()->id;
 
         $client = Client::create($data);
 
@@ -41,6 +50,8 @@ class ClientController extends Controller
      */
     public function show(Client $client)
     {
+        abort_if($client->company_id !== $this->company()->id, 403);
+
         return new ClientResource($client);
     }
 
@@ -49,6 +60,8 @@ class ClientController extends Controller
      */
     public function update(UpdateClientRequest $request, Client $client)
     {
+        abort_if($client->company_id !== $this->company()->id, 403);
+
         $data = $request->validated();
         $client->update($data);
 
@@ -60,6 +73,8 @@ class ClientController extends Controller
      */
     public function destroy(Client $client)
     {
+        abort_if($client->company_id !== $this->company()->id, 403);
+
         $client->delete();
 
         return response('', 204);
